Skv client refactoring (#307)
This PR makes the dto types from the platform and the dto types from skvclient compatible.
* skv::SKVRecord macros are now converted to templates
* skv::FieldType macros are converted to templates
* logging subsystem is refactored to separate subproject
* implement skv::* serialization mechanism which is compatible with the platform
* Use msgpack instead of json for  communication over HTTP
* convert some python skvclient APIs to use msgpack. The rest will be done in a separate PR
* Removed error-prone timestamp constructor

+++ src/k2/tso/service/Service.h
@@ -72,7 +72,7 @@ private: // members
     // is assigned by the CPO at registration time.
     // 0 is a special value which indicates that this TSO worker is currently not registered with the CPO
     // which makes it an instance not authorized to produce timestamps in the cluster.
-    uint64_t _tsoId{0};
+    uint32_t _tsoId{0};
 
     // to poll the gps clock
     std::thread _clockPoller;
